Worked on database backend
I decided to use SQL for the backend.  We can use sqlalchemy which gives
us a good range of potential SQL implemintations that we drop in.  This
will have a bit of a larger initial cost, but in the end it'll be
smoother sailing.

One problem with the current code is that the edit contentent template
uses a normal form to send the data to the backend.  However, the
backend expects a `PUT` request which is only possible if we use ajax to
send the data.  So... we need to figure this out somehow.